Update README.md
Week 2 of the Foundations of Linux in the ALX Program is a crucial step in your journey to mastering the Linux operating system. This week typically focuses on Shell basics, including fundamental command-line skills and navigating the Linux filesystem. Here’s a summary of what you usually learn and do in Week 2: